Simplify `Command` trait
Responsibility for certain functionality has been shifted to `CommandContext`.
The plan is to expand `CommandContext` with functionality that commands may need in future.

This will also allow commands to operate differently on the client and server.
For example, if `CommandContext` is changed to an enum rather than a struct,
it could have a variant for a server context, and one for a client context,
which it's methods would match on to determine how to behave when called by a command.